created the Rectangle class object

+++ csharp/ClassesandObjects/Student.cs
@@ -7,6 +7,6 @@ namespace ClassesandObjects
         public string Course {get; set;}
 
         public int Score {get; set;}
-        public float CGPA {get; set;}
+        public double CGPA {get; set;}
     }
 }
